updated booking

+++ src/app/modules/service/service.service.ts
@@ -3,6 +3,8 @@ import { paginationHelpers } from '../../../helpers/paginationHelper'
 import { IPaginationOptions } from '../../../interfaces/pagination'
 import prisma from '../../../shared/prisma'
 import { serviceSearchableFields } from './service.constant'
+import ApiError from '../../../errors/ApiError'
+import httpStatus from 'http-status'
 
 const createService = async (data: Service) => {
   const result = await prisma.service.create({
@@ -104,6 +106,17 @@ const updateSingleService = async (id: string, payload: Partial<Service>) => {
 }
 
 const deleteSingleService = async (id: string) => {
+  const bookingOrNot = await prisma.booking.findFirst({
+    where: {
+      serviceId: id,
+    },
+  })
+  if (bookingOrNot) {
+    throw new ApiError(
+      httpStatus.BAD_REQUEST,
+      "You can't delete this service, as this service already booking information",
+    )
+  }
   const result = await prisma.service.delete({
     where: {
       id,
